refactor : Winners의 역할을 Cars로 이동 및 Winners 제거
