+++ src/main/java/racingcar/model/Cars.java
@@ -1,6 +1,7 @@
 package racingcar.model;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class Cars {
     private final List<Car> cars;
@@ -15,11 +16,25 @@ public class Cars {
         }
     }
 
-    public Positions getPositions() {
-        return new Positions(cars);
+    public List<Car> getWinnerCars() {
+        int farDistancePosition = farDistancePosition();
+        return findSamePositionCars(farDistancePosition);
     }
 
-    public Winners getWinners() {
-        return new Winners(cars);
+    private int farDistancePosition() {
+        return cars.stream()
+                .mapToInt(Car::getPosition)
+                .max()
+                .orElseThrow();
+    }
+
+    private List<Car> findSamePositionCars(int position) {
+        return cars.stream()
+                .filter(car -> car.getPosition() == position)
+                .collect(Collectors.toList());
+    }
+
+    public Positions getPositions() {
+        return new Positions(cars);
     }
 }
